added new and hopefully better way of splitting spectra prior to centroiding. Todo: Adjust regression to only start at real points

+++ include/qalgorithms_datatypes.h
@@ -85,7 +85,7 @@ namespace qAlgorithms
         // the degrees of freedom are redundant because there are never interpolated points within a centroid
         // we still use them because this way there is no largely redundant second peak fitting function
         std::vector<unsigned int> cumdf; // cumulative df @todo find a better solution for the general case
-        unsigned int startPos, endPos;
+        size_t startPos, endPos;
     };
 
     struct RegCoeffs
